refactor(input): construct state from config
