+++ src/input/mod.rs
@@ -26,9 +26,9 @@ pub use state::{
     BoardPickerCursorHint, ClickHighlightSettings, ColorPickerCursorHint, CommandPaletteCursorHint,
     CommandPaletteState, ContextMenuCursorHint, DrawingState, EyedropperUiState, HelpOverlayClick,
     HelpOverlayCursorHint, HelpOverlayReleaseOutcome, InputHudActiveSource, InputHudEntry,
-    InputHudEntryKind, InputHudSettings, InputState, InputStateSeed, OutputFocusAction,
-    RegionInputSource, RegionPurposeTag, RegionSelectUiState, SelectionHandle,
-    SelectionPropertyEntry, SelectionPropertyKind, TextInputMode, ZoomAction,
+    InputHudEntryKind, InputHudSettings, InputState, OutputFocusAction, RegionInputSource,
+    RegionPurposeTag, RegionSelectUiState, SelectionHandle, SelectionPropertyEntry,
+    SelectionPropertyKind, TextInputMode, ZoomAction,
 };
 #[cfg(feature = "tablet-input")]
 #[allow(unused_imports)]
