+++ src/input/state/core/base/state/mod.rs
@@ -2,6 +2,6 @@ mod init;
 mod modifiers;
 mod structs;
 
-pub use init::InputStateSeed;
+pub(crate) use init::InputStateSeed;
 pub use structs::InputState;
 pub(crate) use structs::{FocusModeRestore, LightModeRestore, PresenterRestore};
